Refactor signature archiving
Tests showed that trying to split up the signature archiving into lots
of small jobs was unreliable so rather than trying to make small jobs
that complete quickly switch to use a long-running job that will
gracefully terminate if the worker shuts down or is restarted.
