+++ app/jobs/archive_petition_job.rb
@@ -104,7 +104,6 @@ class ArchivePetitionJob < ApplicationJob
       end
 
       ArchiveSignaturesJob.perform_later(petition, archived_petition)
-      MarkPetitionAsArchivedJob.perform_later(petition, archived_petition)
     end
   end
