+++ db/migrate/20170710090730_add_archived_at_to_sigantures.rb
@@ -1,0 +1,5 @@
+class AddArchivedAtToSigantures < ActiveRecord::Migration
+  def change
+    add_column :signatures, :archived_at, :datetime
+  end
+end
